IGNITE-22385 rest and cli api to declarative style
