+++ modules/cli/src/integrationTest/java/org/apache/ignite/internal/cli/commands/recovery/partitions/reset/ItResetPartitionsCommandTest.java
@@ -15,15 +15,16 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.cli.commands.recovery;
+package org.apache.ignite.internal.cli.commands.recovery.partitions.reset;
 
-import org.apache.ignite.internal.cli.commands.recovery.reset.ResetPartitionsReplCommand;
+import org.apache.ignite.internal.util.ArrayUtils;
 
-/** Test class for {@link ResetPartitionsReplCommand}. */
-public class ItResetPartitionsReplCommandTest extends ItResetPartitionsTest {
-
+/** Test class for {@link ResetPartitionsCommand}. */
+public class ItResetPartitionsCommandTest extends ItResetPartitionsTest {
     @Override
-    protected Class<?> getCommandClass() {
-        return ResetPartitionsReplCommand.class;
+    protected void execute(String... args) {
+        String[] fullArgs = ArrayUtils.concat(new String[] {"recovery", "partitions", "reset"}, args);
+
+        super.execute(fullArgs);
     }
 }
