+++ modules/cli/src/integrationTest/java/org/apache/ignite/internal/cli/commands/recovery/partitions/restart/ItRestartPartitionsCommandTest.java
@@ -15,16 +15,15 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.cli.commands.recovery.restart;
+package org.apache.ignite.internal.cli.commands.recovery.partitions.restart;
 
 import org.apache.ignite.internal.util.ArrayUtils;
 
 /** Test class for {@link RestartPartitionsCommand}. */
 public class ItRestartPartitionsCommandTest extends ItRestartPartitionsTest {
-
     @Override
     protected void execute(String... args) {
-        String[] fullArgs = ArrayUtils.concat(new String[]{"recovery", "restart-partitions"}, args);
+        String[] fullArgs = ArrayUtils.concat(new String[]{"recovery", "partitions", "restart"}, args);
 
         super.execute(fullArgs);
     }
