+++ modules/cli/src/integrationTest/java/org/apache/ignite/internal/cli/commands/recovery/partitions/states/ItPartitionStatesCommandTest.java
@@ -15,17 +15,15 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.cli.commands.recovery;
+package org.apache.ignite.internal.cli.commands.recovery.partitions.states;
 
-import org.apache.ignite.internal.cli.commands.recovery.reset.ResetPartitionsCommand;
 import org.apache.ignite.internal.util.ArrayUtils;
 
-/** Test class for {@link ResetPartitionsCommand}. */
-public class ItResetPartitionsCommandTest extends ItResetPartitionsTest {
-
+/** Test class for {@link PartitionStatesCommand}. */
+public class ItPartitionStatesCommandTest extends ItPartitionStatesTest {
     @Override
     protected void execute(String... args) {
-        String[] fullArgs = ArrayUtils.concat(new String[] {"recovery", "reset-partitions"}, args);
+        String[] fullArgs = ArrayUtils.concat(new String[] {"recovery", "partitions", "states"}, args);
 
         super.execute(fullArgs);
     }
