+++ modules/cli/src/main/java/org/apache/ignite/internal/cli/commands/recovery/RecoveryReplCommand.java
@@ -18,18 +18,14 @@
 package org.apache.ignite.internal.cli.commands.recovery;
 
 import org.apache.ignite.internal.cli.commands.BaseCommand;
-import org.apache.ignite.internal.cli.commands.recovery.partitions.PartitionStatesReplCommand;
-import org.apache.ignite.internal.cli.commands.recovery.reset.ResetPartitionsReplCommand;
-import org.apache.ignite.internal.cli.commands.recovery.restart.RestartPartitionsReplCommand;
+import org.apache.ignite.internal.cli.commands.recovery.partitions.PartitionsReplCommand;
 import picocli.CommandLine.Command;
 
 /** Disaster recovery command. */
 @Command(name = "recovery",
         subcommands = {
-                PartitionStatesReplCommand.class,
-                ResetPartitionsReplCommand.class,
-                RestartPartitionsReplCommand.class
+                PartitionsReplCommand.class
         },
-        description = "Managers disaster recovery of Ignite cluster")
+        description = "Manages disaster recovery of Ignite cluster")
 public class RecoveryReplCommand extends BaseCommand {
 }
