+++ modules/cli/src/main/java/org/apache/ignite/internal/cli/commands/recovery/partitions/PartitionsCommand.java
@@ -15,17 +15,21 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.cli.commands.recovery;
+package org.apache.ignite.internal.cli.commands.recovery.partitions;
 
 import org.apache.ignite.internal.cli.commands.BaseCommand;
-import org.apache.ignite.internal.cli.commands.recovery.partitions.PartitionsCommand;
+import org.apache.ignite.internal.cli.commands.recovery.partitions.reset.ResetPartitionsCommand;
+import org.apache.ignite.internal.cli.commands.recovery.partitions.restart.RestartPartitionsCommand;
+import org.apache.ignite.internal.cli.commands.recovery.partitions.states.PartitionStatesCommand;
 import picocli.CommandLine.Command;
 
-/** Disaster recovery command. */
-@Command(name = "recovery",
+/** Partitions disaster recovery commands. */
+@Command(name = "partitions",
         subcommands = {
-                PartitionsCommand.class
+                RestartPartitionsCommand.class,
+                ResetPartitionsCommand.class,
+                PartitionStatesCommand.class
         },
-        description = "Manages disaster recovery of Ignite cluster")
-public class RecoveryCommand extends BaseCommand {
+        description = "Manages disaster recovery of Ignite partitions")
+public class PartitionsCommand extends BaseCommand {
 }
