+++ modules/cli/src/main/java/org/apache/ignite/internal/cli/commands/recovery/partitions/restart/RestartPartitionsMixin.java
@@ -15,8 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.cli.commands.recovery.reset;
+package org.apache.ignite.internal.cli.commands.recovery.partitions.restart;
 
+import static org.apache.ignite.internal.cli.commands.Options.Constants.RECOVERY_NODE_NAMES_OPTION;
+import static org.apache.ignite.internal.cli.commands.Options.Constants.RECOVERY_NODE_NAMES_OPTION_DESC;
 import static org.apache.ignite.internal.cli.commands.Options.Constants.RECOVERY_PARTITION_IDS_OPTION;
 import static org.apache.ignite.internal.cli.commands.Options.Constants.RECOVERY_PARTITION_IDS_OPTION_DESC;
 import static org.apache.ignite.internal.cli.commands.Options.Constants.RECOVERY_TABLE_NAME_OPTION;
@@ -25,17 +27,21 @@ import static org.apache.ignite.internal.cli.commands.Options.Constants.RECOVERY
 import static org.apache.ignite.internal.cli.commands.Options.Constants.RECOVERY_ZONE_NAME_OPTION_DESC;
 
 import java.util.List;
+import javax.annotation.Nullable;
 import org.apache.ignite.internal.cli.commands.cluster.ClusterUrlMixin;
 import picocli.CommandLine.Mixin;
 import picocli.CommandLine.Option;
 
-/** Arguments for recovery reset partitions command. */
-public class ResetPartitionsMixin {
+/** Arguments for recovery restart partitions command. */
+public class RestartPartitionsMixin {
     @Mixin
     private ClusterUrlMixin clusterUrl;
 
     @Option(names = RECOVERY_PARTITION_IDS_OPTION, description = RECOVERY_PARTITION_IDS_OPTION_DESC, split = ",")
     private List<Integer> partitionIds;
+
+    @Option(names = RECOVERY_NODE_NAMES_OPTION, description = RECOVERY_NODE_NAMES_OPTION_DESC, split = ",")
+    private List<String> nodeNames;
 
     @Option(names = RECOVERY_ZONE_NAME_OPTION, description = RECOVERY_ZONE_NAME_OPTION_DESC, required = true)
     private String zoneName;
@@ -43,19 +49,26 @@ public class ResetPartitionsMixin {
     @Option(names = RECOVERY_TABLE_NAME_OPTION, description = RECOVERY_TABLE_NAME_OPTION_DESC, required = true)
     private String tableName;
 
-    /** Returns name of the zone to reset partitions of. */
+    /** Returns name of the zone to restart partitions of. */
     public String zoneName() {
         return zoneName;
     }
 
-    /** Returns name of the table to reset partitions of. */
+    /** Returns name of the table to restart partitions of. */
     public String tableName() {
         return tableName;
     }
 
-    /** Returns IDs of partitions to reset. */
+    /** Returns IDs of partitions to restart partitions of. Empty/null means "all partitions". */
+    @Nullable
     public List<Integer> partitionIds() {
         return partitionIds;
+    }
+
+    /** Returns names specifying nodes to restart partitions. Empty/null means "all nodes". */
+    @Nullable
+    public List<String> nodeNames() {
+        return nodeNames;
     }
 
     /** Returns cluster endpoint URL. */
