+++ modules/rest/src/integrationTest/java/org/apache/ignite/internal/rest/recovery/ItDisasterRecoveryControllerResetPartitionsTest.java
@@ -19,6 +19,7 @@ package org.apache.ignite.internal.rest.recovery;
 
 import static org.apache.ignite.internal.TestDefaultProfilesNames.DEFAULT_AIPERSIST_PROFILE_NAME;
 import static org.apache.ignite.internal.rest.constants.HttpCode.OK;
+import static org.apache.ignite.internal.rest.recovery.ItDisasterRecoveryControllerTest.RESET_PARTITIONS_ENDPOINT;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 
@@ -31,14 +32,14 @@ import io.micronaut.test.extensions.junit5.annotation.MicronautTest;
 import jakarta.inject.Inject;
 import java.util.Set;
 import org.apache.ignite.internal.Cluster;
-import org.apache.ignite.internal.ClusterPerTestIntegrationTest;
+import org.apache.ignite.internal.ClusterPerClassIntegrationTest;
 import org.apache.ignite.internal.rest.api.recovery.ResetPartitionsRequest;
-import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 /** Test for disaster recovery reset partitions command, positive cases. */
 @MicronautTest
-public class ItDisasterRecoveryControllerResetPartitionsTest extends ClusterPerTestIntegrationTest {
+public class ItDisasterRecoveryControllerResetPartitionsTest extends ClusterPerClassIntegrationTest {
     private static final String NODE_URL = "http://localhost:" + Cluster.BASE_HTTP_PORT;
 
     private static final String FIRST_ZONE = "first_ZONE";
@@ -51,16 +52,16 @@ public class ItDisasterRecoveryControllerResetPartitionsTest extends ClusterPerT
     @Client(NODE_URL + "/management/v1/recovery/")
     HttpClient client;
 
-    @BeforeEach
+    @BeforeAll
     public void setUp() {
-        executeSql(String.format("CREATE ZONE \"%s\" WITH storage_profiles='%s'", FIRST_ZONE, DEFAULT_AIPERSIST_PROFILE_NAME));
-        executeSql(String.format("CREATE TABLE PUBLIC.\"%s\" (id INT PRIMARY KEY, val INT) WITH PRIMARY_ZONE = '%s'", TABLE_NAME,
+        sql(String.format("CREATE ZONE \"%s\" WITH storage_profiles='%s'", FIRST_ZONE, DEFAULT_AIPERSIST_PROFILE_NAME));
+        sql(String.format("CREATE TABLE PUBLIC.\"%s\" (id INT PRIMARY KEY, val INT) WITH PRIMARY_ZONE = '%s'", TABLE_NAME,
                 FIRST_ZONE));
     }
 
     @Test
     public void testResetAllPartitions() {
-        MutableHttpRequest<ResetPartitionsRequest> post = HttpRequest.POST("/reset-partitions",
+        MutableHttpRequest<ResetPartitionsRequest> post = HttpRequest.POST(RESET_PARTITIONS_ENDPOINT,
                 new ResetPartitionsRequest(FIRST_ZONE, QUALIFIED_TABLE_NAME, Set.of()));
 
         HttpResponse<Void> response = client.toBlocking().exchange(post);
@@ -70,7 +71,7 @@ public class ItDisasterRecoveryControllerResetPartitionsTest extends ClusterPerT
 
     @Test
     public void testResetSpecifiedPartitions() {
-        MutableHttpRequest<ResetPartitionsRequest> post = HttpRequest.POST("/reset-partitions",
+        MutableHttpRequest<ResetPartitionsRequest> post = HttpRequest.POST(RESET_PARTITIONS_ENDPOINT,
                 new ResetPartitionsRequest(FIRST_ZONE, QUALIFIED_TABLE_NAME, Set.of(0)));
 
         HttpResponse<Void> response = client.toBlocking().exchange(post);
